# swagger/swagger.yaml
openapi: 3.0.0
info:
  description: API онлайн-аукциона
  version: '1.0'
  title: Auction API, Tinkoff Fintech School
servers:
  - url: '/v1/auction'
paths:
...
  /users/{id}:
    get:
      summary: Получить пользователя
      description: >
        Метод возвращает информацию о пользователе.
        При id = 0 возвращается информация о пользователе, который выполняет запрос.
      operationId: GetUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя. 0 - текущий пользователь
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Успешный ответ с сущностью пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Обновить информацию о пользователе
      description: >
        Обновлять можно информацию только о текущем пользователе, т.е. id = 0.
        Обновление email-а, пароля и системных полей не допускается
      operationId: UpdateUser
      tags: [users]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Идентификатор пользователя
          schema:
            type: integer
            format: int64
            enum: [0]
          required: true
      requestBody:
        description: Сущность обновлённого пользователя
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: Имя
                  example: Павел
                last_name:
                  type: string
                  description: Фамилия
                  example: Дуров
                birthday:
                  type: string
                  format: date
                  description: Дата рождения
                  example: "1984-10-10"
                  default: ""
      responses:
        '200':
          description: Успешный ответ с сущностью пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Неверные входные данные
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Контент по переданному идентификатору не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Неавторизованный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор пользователя
          example: 1
        first_name:
          type: string
          description: Имя
          example: Павел
        last_name:
          type: string
          description: Фамилия
          example: Дуров
        birthday:
          type: string
          format: date
          description: Дата рождения
          example: "1984-10-10"
          default: ""
        email:
          type: string
          format: email
          description: Email, доступен только его владельцу
          example: durov@telegram.org
        created_at:
          type: string
          format: date-time
...